Separate the environment and backend tests
